+++ rollup.config.prd.mjs
@@ -1,13 +1,14 @@
 // rollup.config.prd.mjs
-import baseConfig from './rollup.config.base.mjs';
+import { createBaseConfig } from './rollup.config.base.mjs';
 import del from 'rollup-plugin-delete';
 import terser from '@rollup/plugin-terser';
+
+const baseConfig = createBaseConfig({ sourceMap: false });
 
 export default {
   ...baseConfig,
   output: {
     ...baseConfig.output,
-    sourcemap: false,
   },
   plugins: [del({ targets: 'dist/**' }), ...baseConfig.plugins, terser()],
 };